Feat : 재고상ì 재고 상세 보여줄때 기존 보유 재고, ê 가져오기

+++ src/main/java/com/seahere/backend/inventory/entity/InventoryEntity.java
@@ -28,7 +28,7 @@ public class InventoryEntity {
     private CompanyEntity company;
 
     @Column(name = "quantity")
-    private int quantity;
+    private float quantity;
 
     @Column(name = "category")
     private String category;
